backend/requirements.txt
flask[async]
flask-cors
google-generativeai
python-dotenv 